requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
phonenumbers==9.0.4
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
